fixed some things

+++ ISD93xx_NAU85L40_1C_MIC_48kHz_DataOut/ISD93xx_NAU85L40_1C_MIC/ISD93xx_NAU85L40_1C/isd9340_keil/sample/driver/smpl_drvi2s_nau85L40_1C/Beamforming.c
@@ -63,7 +63,7 @@ int main(int argc, char* argv[]){
 	//run(in);
 }
 
-uint32_t run(float in0, float in1, float in2, float in3){
+uint16_t run(float in0, float in1, float in2, float in3){
 	float output = 0;
 	float sum;
 	sum = delay_out(0, in0);
